% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{basemap}
\alias{basemap}
\title{A simple plotting function for seafloor climate projections}
\usage{
basemap(r, title = names(r), colours = NULL, ranges = c(0, 1))
}
\arguments{
\item{r}{A Raster or RasterBrick object}

\item{colours}{A vector of colors to use for the color key}

\item{ranges}{A numeric vector of length two providing quantile limits for the color key scale}

\item{titles}{A character vector string of the map titles}
}
\value{
This function returns a base ggplot object, so other ggplot components can be added.
}
\description{
This function is a wrapper of  \code{\link{ggplot2}}
}
\details{
This is a simple map function to visualize the ensemble average of seafloor climate projections.
The Taiwan EEZ boundary is downloaded from \url{https://marineregions.org/gazetteer.php?p=details&id=8321}.
The 200-m (dashed line) and 4000-m depth (solid line) contours were drawn using bathymetric data from the
\code{\link{etopo2022}}.
}
\examples{
basemap(-etopo2022, title="Depth (m)", colours = terrain.colors(7))

}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
