% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_1950_2000_sd.R
\docType{data}
\name{cmip6_1950_2000_sd}
\alias{cmip6_1950_2000_sd}
\title{Model-Averaging Decadal Standard deviation of Seafloor Climatic Data during 1950 to 2000}
\format{
A RasterBrick object of 11 raster layers:
\describe{
  \item{epc_sd_1950_to_2000}{Export POC flux to seafloor (mg C m^-2 d^-1)}
  \item{o2_sd_1950_to_2000}{Dissolved oxygen concentration at seafloor (mol m^-3)}
  \item{ph_sd_1950_to_2000}{pH at seafloor}
  \item{thetao_sd_1950_to_2000}{Potential temperature at seafllor (C)}
  \item{arag_sd_1950_to_2000}{Aragonite Concentration (mol m-3)}
  \item{calc_sd_1950_to_2000}{Calcite Concentration (mol m-3)}
  \item{co3_sd_1950_to_2000}{Mole Concentration of Carbonate expressed as Carbon in Sea Water (mol m-3)}
  \item{co3satarag_sd_1950_to_2000}{Carbonate ion concentration for seawater in equilibrium with pure aragonite (mol m-3)}
  \item{co3satcalc_sd_1950_to_2000}{Carbonate ion concentration for seawater in equilibrium with pure calcite (mol m-3)}
  \item{aragsat_sd_1950_to_2000}{Aragonite Saturation State (omega)}
  \item{calcsat_sd_1950_to_2000}{Calcite Saturation State (omega)}
}
}
\source{
\url{https://esgf-node.llnl.gov/search/esgf-llnl/}
}
\description{
Emsemble average of the decadal Standard deviation of seafloor (bottom-most) projections from 1950 to 2000.
}
\details{
Yearly mean projections from 1950 to 2000 were calculated from each models.
Decadal standard deviation was calculated from the yearly mean and then averaged across all models.
}
\examples{
plotmap(subset(cmip6_1950_2000_sd, 1:4))
}
\keyword{datasets}
