% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_1950_2014_voccMeg.R
\docType{data}
\name{cmip6_1950_2014_voccMeg}
\alias{cmip6_1950_2014_voccMeg}
\title{Model-Averaging of Gradient-based climate velocity magnitude during 1950 to 2014}
\format{
A RasterBrick object of 11 raster layers:
\describe{
  \item{epc_voccAng_1950_to_2014}{Export POC flux to seafloor (km/yr)}
  \item{o2_voccAng_1950_to_2014}{Dissolved oxygen concentration at seafloor (km/yr)}
  \item{ph_voccAng_1950_to_2014}{pH at seafloor (km/yr)}
  \item{thetao_voccAng_1950_to_2014}{Potential temperature at seafllor (km/yr)}
  \item{arag_voccAng_1950_to_2014}{Aragonite Concentration (km/yr)}
  \item{calc_voccAng_1950_to_2014}{Calcite Concentration (km/yr)}
  \item{co3_voccAng_1950_to_2014}{Mole Concentration of Carbonate expressed as Carbon in Sea Water (km/yr)}
  \item{co3satarag_voccAng_1950_to_2014}{Carbonate ion concentration for seawater in equilibrium with pure aragonite (km/yr)}
  \item{co3satcalc_voccAng_1950_to_2014}{Carbotoponate ion concentration for seawater in equilibrium with pure calcite (km/yr)}
  \item{aragsat_voccAng_1950_to_2014}{Aragonite saturation State (km/yr)}
  \item{calcsat_voccAng_1950_to_2014}{Calcite saturation State (km/yr)}
}
}
\source{
\url{https://esgf-node.llnl.gov/search/esgf-llnl/}
}
\description{
Ensemble average of the seafloor gradient-based climate velocity magnitudes during 1950 to 2014.
}
\details{
Gradient-based climate velocity magnitude from 1950 to 2014 were calculated from each models and then averaged across all models.
Climate velocity was calculated using package gVoCC \url{https://rdrr.io/github/JorGarMol/VoCC/man/gVoCC.html}.
}
\examples{
plotmap(subset(cmip6_1950_2014_voccMeg, 1:4))
}
\keyword{datasets}
