% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_2041_2060_ch.R
\docType{data}
\name{cmip6_2041_2060_ch}
\alias{cmip6_2041_2060_ch}
\title{Seafloor Climate Change by 2041 to 2060}
\format{
A RasterBrick object of 11 raster layers:
\describe{
  \item{epc_av_2041_to_2060_minus_1950_to_2000}{Export POC flux to seafloor (mg C m^-2 d^-1)}
  \item{o2_av_2041_to_2060_minus_1950_to_2000}{Dissolved oxygen concentration at seafloor (mol m^-3)}
  \item{ph_av_2041_to_2060_minus_1950_to_2000}{Difference in pH at seafloor}
  \item{thetao_av_2041_to_2060_minus_1950_to_2000}{Potential temperature at seafllor (C)}
  \item{arag_av_2041_to_2060_minus_1950_to_2000}{Aragonite concentration (mol m-3)}
  \item{calc_av_2041_to_2060_minus_1950_to_2000}{Calcite concentration (mol m-3)}
  \item{co3_av_2041_to_2060_minus_1950_to_2000}{Mole concentration of Carbonate expressed as Carbon in Sea Water (mol m-3)}
  \item{co3satarag_av_2041_to_2060_minus_1950_to_2000}{Carbonate ion concentration for seawater in equilibrium with pure aragonite (mol m-3)}
  \item{co3satcalc_av_2041_to_2060_minus_1950_to_2000}{Carbonate ion concentration for seawater in equilibrium with pure calcite (mol m-3)}
  \item{aragsat_av_2041_to_2060_minus_1950_to_2000}{Aragonite Saturation State (omega)}
  \item{calcsat_av_2041_to_2060_minus_1950_to_2000}{Calcite Saturation State (omega)}
}
}
\source{
\url{https://esgf-node.llnl.gov/search/esgf-llnl/}
}
\description{
Emsemble average of the decadal mean seafloor (bottom-most) projections during 2041 to 2060
minus the decadal mean projections during 1950 to 2000.
}
\details{
Climate change (difference of decadal mean) between the period of 1950 to 2000 and the period 2041 to 2080 were calculated
for each models. Emsemble average of climate changes was calculated across all models.
}
\examples{
library(ggplot2)
library(dplyr)
library(patchwork)
library(tidyr)
library(sf)

# Convert raster to data frame and then to list
cmip6_list <- subset(cmip6_2041_2060_ch, 1:4) \%>\%
  as.data.frame(xy = TRUE) \%>\%
  gather(-x, -y, key = "var", value = "value") \%>\%
  group_split(var)

# Depth
bathy <- etopo2022\%>\% as.data.frame(xy = TRUE) \%>\% na.omit

# ggolot list
gg_list = lapply(cmip6_list, function(dat) {
  lim <- max(abs(quantile(dat$value, c(0.01, 0.99), na.rm=TRUE)))
  ggplot(dat) +
    geom_raster(aes(x=x, y=y, fill=value))+
    geom_polygon(data=land, aes(x=X, y=Y, group=PID), fill="bisque2", colour="black")+
    geom_sf(data=as(eez, "sf"), fill="transparent", colour="red")+
    geom_contour(data=bathy, aes(x=x, y=y, z=layer), breaks=-200, linetype=2, colour="gray50")+
    geom_contour(data=bathy, aes(x=x, y=y, z=layer), breaks=seq(-6000, -1000, by=1000), linetype=1, colour="gray50")+
    scale_fill_gradientn(colours = jet.colors3(10), limits=c(-lim, lim))+
    scale_x_continuous(expand = expansion(mult = 0))+
    scale_y_continuous(expand = expansion(mult = 0))+
    labs(x="Longitude", y="Latitude", fill="")+
    facet_wrap(~ var) +
    theme_bw()
    })

# Wrap ggplot list
wrap_plots(gg_list)
}
\keyword{datasets}
