% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_2041_2060_sd.R
\docType{data}
\name{cmip6_2041_2060_sd}
\alias{cmip6_2041_2060_sd}
\title{Model-Averaging Decadal Standard deviation of Seafloor Climatic Data during 2041 to 2060}
\format{
A RasterBrick object of 11 raster layers:
\describe{
  \item{epc_sd_2041_to_2060}{Export POC flux to seafloor (mg C m^-2 d^-1)}
  \item{o2_sd_2041_to_2060}{Dissolved oxygen concentration at seafloor (mol m^-3)}
  \item{ph_sd_2041_to_2060}{pH at seafloor}
  \item{thetao_sd_2041_to_2060}{Potential temperature at seafllor (C)}
  \item{arag_sd_2041_to_2060}{Aragonite Concentration (mol m-3)}
  \item{calc_sd_2041_to_2060}{Calcite Concentration (mol m-3)}
  \item{co3_sd_2041_to_2060}{Mole Concentration of Carbonate expressed as Carbon in Sea Water (mol m-3)}
  \item{co3satarag_sd_2041_to_2060}{Carbonate ion concentration for seawater in equilibrium with pure aragonite (mol m-3)}
  \item{co3satcalc_sd_2041_to_2060}{Carbonate ion concentration for seawater in equilibrium with pure calcite (mol m-3)}
  \item{aragsat_sd_2041_to_2060}{Aragonite Saturation State (omega)}
  \item{calcsat_sd_2041_to_2060}{Calcite Saturation State (omega)}
}
}
\source{
\url{https://esgf-node.llnl.gov/search/esgf-llnl/}
}
\description{
Emsemble average of the decadal Standard deviation of seafloor (bottom-most) projections from 2041 to 2060.
}
\details{
Yearly mean projections from 2041 to 2060 were calculated from each models.
Decadal standard deviation was calculated from the yearly mean and then averaged across all models.
}
\keyword{datasets}
