% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_2081_2100_exsd.R
\docType{data}
\name{cmip6_2081_2100_exsd}
\alias{cmip6_2081_2100_exsd}
\title{Exposure to Climate Change Hazard during 2081 to 2100}
\format{
A RasterBrick object of 11 raster layers:
\describe{
  \item{epc_exsd_2081_to_2100}{Exposure to climate change hazard in export POC flux to seafloor}
  \item{o2_exsd_2081_to_2100}{Exposure to climate change hazard in dissolved oxygen concentration at seafloor}
  \item{ph_exsd_2081_to_2100}{Exposure to climate change hazard in pH at seafloor}
  \item{thetao_exsd_2081_to_2100}{Exposure to climate change hazard in potential temperature at seafllor}
  \item{arag_exsd_2081_to_2100}{Exposure to climate change hazard in aragonite Concentration}
  \item{calc_exsd_2081_to_2100}{Exposure to climate change hazard in calcite Concentration}
  \item{co3_exsd_2081_to_2100}{Exposure to climate change hazard in mole Concentration of Carbonate expressed as Carbon in Sea Water}
  \item{co3satarag_exsd_2081_to_2100}{Exposure to climate change hazard in carbonate ion concentration for seawater in equilibrium with pure aragonite}
  \item{co3satcalc_exsd_2081_to_2100}{Exposure to climate change hazard in carbonate ion concentration for seawater in equilibrium with pure calcite}
  \item{aragsat_exsd_2081_to_2100}{Exposure to climate change hazard in aragonite Saturation State}
  \item{calcsat_exsd_2081_to_2100}{Exposure to climate change hazard in calcite Saturation State}
}
}
\source{
\url{https://esgf-node.llnl.gov/search/esgf-llnl/}
}
\description{
Magnitude of climate change by 2081 to 2100 in the unit of historical variability
}
\details{
Exposure to climate change hazard was calculated as \code{\link{cmip6_2081_2100_ch}} divided by \code{\link{cmip6_1950_2000_sd}}.
}
\keyword{datasets}
