% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral.R
\docType{data}
\name{coral}
\alias{coral}
\title{Global distribution of cold-water corals}
\format{
A SpatialPointsDataFrame
}
\source{
\url{http://data.unep-wcmc.org/datasets/3}
}
\description{
Global distribution of cold-water corals from Freiwald et al. (2017)
}
\details{
This dataset has compiled occurrence records of cold-water corals in the Argentina EEZ.
}
\examples{
basemap(-etopo2022, title="Depth (m)", colours = terrain.colors(7))+
geom_sf(data=as(coral, "sf"), size=0.8)
}
\references{
\itemize{
  \item{Freiwald A, Rogers A, Hall-Spencer J, Guinotte JM, Davies AJ, Yesson C, Martin CS, Weatherdon LV (2021). Global distribution of cold-water corals (version 5.1).}
  \item{Fifth update to the dataset in Freiwald et al. (2004) by UNEP-WCMC, in collaboration with Andre Freiwald and John Guinotte. Cambridge (UK): UN Environment Programme World Conservation Monitoring Centre. Data DOI: \url{https://doi.org/10.34892/72×9-rt61}}
  \item{Freiwald A, Fosså JH, Grehan A, Koslow T, Roberts JM (2004). Cold-water coral reefs: out of sight – no longer out of mind. Biodiversity Series 22. Cambridge (UK): UNEP World Conservation Monitoring Centre. 86 pp. URL: \url{https://archive.org/details/coldwatercoralre04frei}}
  \item{OSPAR Commission. (2015) OSPAR Threatened and/or Declining Habitats 2015. URL: \url{http://www.ospar.org/work-areas/bdc/species-habitats/list-of-threatened-declining-species-habitats}. Data URL: \url{http://www.emodnet-seabedhabitats.eu/download}.}
}
}
\keyword{datasets}
