% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo2022.R
\docType{data}
\name{etopo2022}
\alias{etopo2022}
\title{ETOPO2022 Global Relief Model}
\format{
A RasterLayer object of water depth
}
\source{
\url{https://www.ncei.noaa.gov/products/etopo-global-relief-model}
}
\description{
A RasterBrick of the Global Relief Model. A multi-layer raster object compiled from the Cell/pixel-registered ETOPO1 bedrock global relief model.
The grid resolution is reduced from 1 minutes to 0.25 degree by averaging.
}
\examples{
library(ggplot2)
library(dplyr)
library(sf)

bathy <- etopo2022 \%>\% as.data.frame(xy = TRUE) \%>\% na.omit
ggplot(bathy) +
  geom_raster(aes(x=x, y=y, fill=-layer))+
  geom_polygon(data=land, aes(x=X, y=Y, group=PID), fill="bisque2", colour="transparent")+
  geom_sf(data=as(eez, "sf"), fill="transparent", colour="red")+
  geom_contour(data=bathy, aes(x=x, y=y, z=layer), breaks=-200, linetype=2, colour="gray50")+
  geom_contour(data=bathy, aes(x=x, y=y, z=layer), breaks=-4000, linetype=1, colour="gray50")+
  scale_fill_gradientn(colours=terrain.colors(7))+
  scale_x_continuous(expand = expansion(mult = 0))+
  scale_y_continuous(expand = expansion(mult = 0))+
  labs(x=NULL, y=NULL, fill="Depth\n(m)")+
  theme_bw() \%+replace\% theme(legend.position = "right", legend.key.width =  unit(0.5, 'cm'))
}
\keyword{datasets}
