% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.colors.R
\name{jet.colors}
\alias{jet.colors}
\alias{jet.col.rev}
\alias{jet.colors2}
\alias{jet.colors3}
\alias{jet.colors4}
\title{'jet.colors' function}
\usage{
jet.colors(n)

jet.col.rev(n)

jet.colors2(n)

jet.colors3(n)

jet.colors4(n)

jet.colors(n)
}
\arguments{
\item{n}{An integer argument (the required number of colors)}
}
\value{
a character vector of interpolated colors
}
\description{
This function is a wrapper of color interpolation  \code{\link{colorRampPalette}},
which takes an integer argument (the required number of colors)
and returns a character vector of colors
}
\section{Functions}{
\itemize{
\item \code{jet.col.rev()}: Reversed 'jet.colors'

\item \code{jet.colors2()}: Easier-on-eye version of 'jet.colors'

\item \code{jet.colors3()}: same jet.colors' with "white" color in the middle

\item \code{jet.colors4()}: same jet.colors' starts with "white" color

\item \code{jet.colors()}: Same jet.colors in log scale

}}
\examples{
jet.colors(10)
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
