% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.R
\name{plotmap}
\alias{plotmap}
\title{A simple plotting function for multi-panel maps with independent color keys}
\usage{
plotmap(r, titles = names(r), colours = NULL, ranges = c(0, 1))
}
\arguments{
\item{r}{A Raster or RasterBrick object}

\item{titles}{A character vector string of the map titles}

\item{colours}{A vector of colors to use for the color key}

\item{ranges}{A numeric vector of length two providing quantile limits for the color key scale}
}
\description{
This function is a wrapper of  \code{\link{ggplot2}} and \code{\link[patchwork]{wrap_plots}}
}
\details{
This is a simple map function to visualize multiple ensemble average of seafloor climate projections.
First, the RasterBrick containing climate projections is subseted and converted into a data frame. Next, we
stack the data frame using the \code{\link[tidyr]{gather}} and \code{\link[tidyr]{group_split}} into a list.
The Taiwan EEZ boundary is downloaded from \url{https://marineregions.org/gazetteer.php?p=details&id=8321}.
The 200-m (dashed line) and 4000-m depth (solid line) contours were drawn using bathymetric data from the
\code{\link{etopo2022}}. We create multipanel maps using \code{\link{ggplot2}} with each panel corresponding
to a variable to ensure independent color keys. Finally, we wrap the ggplot list using \code{\link[patchwork]{wrap_plots}}.
}
\examples{
plotmap(cmip6_1950_2000_av \%>\% subset(1:4))

}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
