% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vent.R
\docType{data}
\name{vent}
\alias{vent}
\title{Global Distribution of Hydrothermal Vents}
\format{
A SpatialPointsDataFrame
}
\source{
\url{https://doi.pangaea.de/10.1594/PANGAEA.917894}
}
\description{
Global Distribution of Hydrothermal Vent Fields (2020)- InterRidge Global Database of Active Submarine Hydrothermal Vent Fields 3.4
}
\details{
This dataset has compiled 721  separate hythrothermal vent fields in the world ocean.
}
\examples{
basemap(-etopo2022, title="Depth (m)", colours = terrain.colors(7))+
geom_sf(data=as(vent, "sf"), size=0.8)
}
\references{
Beaulieu, Stace E; Szafrański, Kamil M (2020): InterRidge Global Database of Active Submarine Hydrothermal Vent Fields Version 3.4 [dataset]. PANGAE.
}
\keyword{datasets}
