% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canyon.R
\docType{data}
\name{canyon}
\alias{canyon}
\title{Global distribution of large submarine canyons}
\format{
A SpatialLinesDataFrame
}
\source{
\url{https://doi.org/10.1016/j.margeo.2011.05.008}
}
\description{
Global distribution of large submarine canyons from Harris and Whiteway (2011)
}
\details{
This dataset has compiled 5849 separate large submarine canyons in the world ocean.
}
\examples{
basemap(-etopo2022, title="Depth (m)", colours = terrain.colors(7))+
geom_sf(data=as(canyon, "sf"), size=0.8)
}
\references{
Harris, P.T., Whiteway, T., 2011. Global distribution of large submarine canyons:
Geomorphic differences between active and passive continental margins. Marine Geology 285, 69–86.
}
\keyword{datasets}
