% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_2081_2100_voccAng.R
\docType{data}
\name{cmip6_2081_2100_voccAng}
\alias{cmip6_2081_2100_voccAng}
\title{Model-Averaging of Gradient-based climate velocity angle during 2081 to 2100}
\format{
A RasterBrick object of 11 raster layers:
\describe{
  \item{epc_voccAng_2081_to_2100}{Export POC flux to seafloor (degree)}
  \item{o2_voccAng_2081_to_2100}{Dissolved oxygen concentration at seafloor (degree)}
  \item{ph_voccAng_2081_to_2100}{pH at seafloor (degree)}
  \item{thetao_voccAng_2081_to_2100}{Potential temperature at seafllor (degree)}
  \item{arag_voccAng_2081_to_2100}{Aragonite Concentration (degree)}
  \item{calc_voccAng_2081_to_2100}{Calcite Concentration (degree)}
  \item{co3_voccAng_2081_to_2100}{Mole Concentration of Carbonate expressed as Carbon in Sea Water (degree)}
  \item{co3satarag_voccAng_2081_to_2100}{Carbonate ion concentration for seawater in equilibrium with pure aragonite (degree)}
  \item{co3satcalc_voccAng_2081_to_2100}{Carbonate ion concentration for seawater in equilibrium with pure calcite (degree)}
  \item{aragsat_voccAng_2081_to_2100}{Aragonite saturation State (degree)}
  \item{calcsat_voccAng_2081_to_2100}{Calcite saturation State (degree)}
}
}
\source{
\url{https://esgf-node.llnl.gov/search/esgf-llnl/}
}
\description{
Ensemble average of the seafloor gradient-based climate velocity angle during 2081 to 2100.
}
\details{
Gradient-based climate velocity angle from 2081 to 2100 were calculated from each models and then averaged across all models.
Climate velocity was calculated using package gVoCC \url{https://rdrr.io/github/JorGarMol/VoCC/man/gVoCC.html}.
}
\examples{
library(ggplot2)
library(dplyr)
library(patchwork)
library(tidyr)
library(sf)

# Convert raster to data frame and then to list
cmip6_list <- subset(cmip6_2081_2100_voccAng, 1:4) \%>\%
  as.data.frame(xy = TRUE) \%>\%
  gather(-x, -y, key = "var", value = "value") \%>\%
  group_split(var)

# Depth
bathy <- etopo2022\%>\% as.data.frame(xy = TRUE) \%>\% na.omit

# ggolot list
gg_list = lapply(cmip6_list, function(dat) {
  ggplot(dat) +
    geom_raster(aes(x=x, y=y, fill=value))+
    geom_polygon(data=land, aes(x=X, y=Y, group=PID), fill="bisque2", colour="black")+
    geom_sf(data=as(eez, "sf"), fill="transparent", colour="red")+
    geom_contour(data=bathy, aes(x=x, y=y, z=layer), breaks=-200, linetype=2, colour="gray50")+
    geom_contour(data=bathy, aes(x=x, y=y, z=layer), breaks=seq(-6000, -1000, by=1000), linetype=1, colour="gray50")+
    scale_fill_gradientn(colours = jet.colors(10))+
    scale_x_continuous(expand = expansion(mult = 0))+
    scale_y_continuous(expand = expansion(mult = 0))+
    labs(x="Longitude", y="Latitude", fill="")+
    facet_wrap(~ var) +
    theme_bw()
    })

# Wrap ggplot list
wrap_plots(gg_list)
}
\keyword{datasets}
