% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_2081_2100_voccMeg.R
\docType{data}
\name{cmip6_2081_2100_voccMeg}
\alias{cmip6_2081_2100_voccMeg}
\title{Model-Averaging of Gradient-based climate velocity magnitude during 2081 to 2100}
\format{
A RasterBrick object of 11 raster layers:
\describe{
  \item{epc_voccAng_2081_to_2100}{Export POC flux to seafloor (km/yr)}
  \item{o2_voccAng_2081_to_2100}{Dissolved oxygen concentration at seafloor (km/yr)}
  \item{ph_voccAng_2081_to_2100}{pH at seafloor (km/yr)}
  \item{thetao_voccAng_2081_to_2100}{Potential temperature at seafllor (km/yr)}
  \item{arag_voccAng_2081_to_2100}{Aragonite Concentration (km/yr)}
  \item{calc_voccAng_2081_to_2100}{Calcite Concentration (km/yr)}
  \item{co3_voccAng_2081_to_2100}{Mole Concentration of Carbonate expressed as Carbon in Sea Water (km/yr)}
  \item{co3satarag_voccAng_2081_to_2100}{Carbonate ion concentration for seawater in equilibrium with pure aragonite (km/yr)}
  \item{co3satcalc_voccAng_2081_to_2100}{Carbonate ion concentration for seawater in equilibrium with pure calcite (km/yr)}
  \item{aragsat_voccAng_2081_to_2100}{Aragonite saturation State (km/yr)}
  \item{calcsat_voccAng_2081_to_2100}{Calcite saturation State (km/yr)}
}
}
\source{
\url{https://esgf-node.llnl.gov/search/esgf-llnl/}
}
\description{
Ensemble average of the seafloor gradient-based climate velocity magnitudes during 2081 to 2100.
}
\details{
Gradient-based climate velocity magnitude from 2081 to 2100 were calculated from each models and then averaged across all models.
Climate velocity was calculated using package gVoCC \url{https://rdrr.io/github/JorGarMol/VoCC/man/gVoCC.html}.
}
\examples{
plotmap(subset(cmip6_2081_2100_voccMeg, 1:4))
}
\keyword{datasets}
