% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo2022.R
\docType{data}
\name{etopo2022}
\alias{etopo2022}
\title{ETOPO2022 Global Relief Model}
\format{
A RasterLayer object of water depth
}
\source{
\url{https://www.ncei.noaa.gov/products/etopo-global-relief-model}
}
\description{
A RasterBrick of the Global Relief Model. A multi-layer raster object compiled from the Cell/pixel-registered ETOPO1 bedrock global relief model.
The grid resolution is reduced from 1 minutes to 0.25 degree by averaging.
}
\examples{
plotmap(-etopo2022, colours=terrain.colors(7), titles="Depth (m)")
}
\keyword{datasets}
