% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{The Fish Database of Taiwan}
\format{
A SpatialPointsDataFrame
}
\source{
\url{https://www.gbif.org/dataset/b7dd8170-98a0-11de-b4d9-b8a03c50a862}
}
\description{
The Fish Database of Taiwan
}
\details{
This dataset has compiled 24955 fish occurrence data in Taiwan.
}
\examples{
basemap(-etopo2022, title="Depth (m)", colours = terrain.colors(7))+
geom_sf(data=as(fish, "sf"), size=0.8)
}
\references{
GBIF.org (10 September 2024) GBIF Occurrence Download \url{https://doi.org/10.15468/dl.e57g23}
}
\keyword{datasets}
